+++ src/main/java/com/telegrambot/cache/UserDataCache.java
@@ -8,15 +8,15 @@ import java.util.Map;
 
 @Service
 public class UserDataCache implements DataCache {
-    private Map<Integer, BotState> usersBotStates = new HashMap<>();
+    private Map<Long, BotState> usersBotStates = new HashMap<>();
 
     @Override
-    public void setUsersCurrentBotState(int userId, BotState botState) {
+    public void setUsersCurrentBotState(Long userId, BotState botState) {
         usersBotStates.put(userId, botState);
     }
 
     @Override
-    public BotState getUsersCurrentBotState(int userId) {
+    public BotState getUsersCurrentBotState(Long userId) {
         BotState botState = usersBotStates.get(userId);
         if (botState == null) {
             botState = BotState.SHOW_MAIN_MENU;
